+++ sdmx/source/__init__.py
@@ -21,6 +21,7 @@ DataContentType = Enum("DataContentType", "CSV JSON XML")
 #: endpoints that are described in the standards but are not implemented by any source
 #: currently in :file:`sources.json`; these all return 404.
 SDMX_ML_SUPPORTS = {
+    Resource.availableconstraint: False,
     Resource.attachementconstraint: False,
     Resource.customtypescheme: False,
     Resource.data: True,